+++ tacocloud-messaging-jms/src/main/java/tacos/send/JmsOrderMessagingService.java
@@ -1,8 +1,9 @@
-package tacos;
+package tacos.send;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jms.core.JmsTemplate;
 import org.springframework.stereotype.Service;
+import tacos.Order;
 
 import javax.jms.JMSException;
 import javax.jms.Message;
