+++ tacocloud-messaging-jms/src/main/java/tacos/send/OrderMessagingService.java
@@ -1,4 +1,6 @@
-package tacos;
+package tacos.send;
+
+import tacos.Order;
 
 public interface OrderMessagingService {
